# docs/API.openapi.yaml
openapi: 3.0.0
info:
  title: BuildBrainOS API
  description: |
    Complete REST API specification for BuildBrainOS platform.
    All endpoints require JWT authentication in Authorization header.
    
    Base URLs:
    - Development: http://localhost:4000/api/v1
    - Production: https://api.buildbrainos.com/api/v1
    
    Authentication:
    - JWT Bearer Token in Authorization header
    - Refresh tokens valid for 30 days
    - Rate limiting: 1000 requests/hour per user
  version: 1.0.0
  contact:
    name: BuildBrainOS Support
    email: support@buildbrainos.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.buildbrainos.com/api/v1
    description: Production server
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - role
                - companyName
              properties:
                email:
                  type: string
                  format: email
                  example: gc@constructionco.com
                password:
                  type: string
                  minLength: 8
                  example: SecurePass123!
                role:
                  type: string
                  enum: [GC, SUBCONTRACTOR, WORKER, GOVERNMENT, SUPPLIER]
                  example: GC
                companyName:
                  type: string
                  example: BuildCorp Solutions
                phone:
                  type: string
                  example: +1-555-123-4567
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/kyc/start:
    post:
      tags:
        - Authentication
      summary: Start KYC verification
      operationId: startKYC
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tier
              properties:
                tier:
                  type: string
                  enum: [TIER_1, TIER_2, TIER_3]
                  description: KYC verification tier
      responses:
        '200':
          description: KYC process initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  kycInquiryId:
                    type: string
                  redirectUrl:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
...
  /documents/upload:
    post:
      tags:
        - Documents
      summary: Upload project document
      operationId: uploadDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - type
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [CONTRACT, LIEN_WAIVER, INSURANCE, LICENSE, BLUEPRINT, PERMIT, RFI]
                projectId:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [GC, SUBCONTRACTOR, WORKER, GOVERNMENT, SUPPLIER]
        companyName:
          type: string
        kycTier:
          type: string
          enum: [TIER_1, TIER_2, TIER_3]
        kycStatus:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        profile:
          type: object
          properties:
            phone:
              type: string
            bio:
              type: string
            avatar:
              type: string
              format: uri
            address:
              $ref: '#/components/schemas/Address'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (expires in 7 days)
        refreshToken:
          type: string
          description: Refresh token (expires in 30 days)
        user:
          $ref: '#/components/schemas/User'
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
...
    Document:
      type: object
      properties:
        id:
          type: string
        fileName:
          type: string
        type:
          type: string
          enum: [CONTRACT, LIEN_WAIVER, INSURANCE, LICENSE, BLUEPRINT, PERMIT, RFI]
        url:
          type: string
          format: uri
        status:
          type: string
          enum: [UPLOADED, PROCESSING, VERIFIED, FAILED]
        extractedData:
          type: object
        aiVerified:
          type: boolean
        uploadedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
...